Code refactored - proper module exports
