+++ css_tasks/task11/js/elements.js
@@ -1,11 +1,4 @@
-// export const DOMElements = {
-//   gallery: document.querySelector('.gallery'),
-//   popup: document.querySelector('.popup'),
-//   body: document.querySelector('body'),
-//   overlay: document.querySelector('.overlay'),
-// };
-
-export const resolver = new Promise((res) => {
+export const elements = new Promise((res) => {
   document.addEventListener('DOMContentLoaded', () => {
     const DOMElements = {
       gallery: document.querySelector('.gallery'),
